# specs/001-architect-client-portal/contracts/api.yaml
openapi: "3.1.0"
info:
  title: ArchiDrive API
  version: "1.0.0"
  description: Architect-Client Design Portal — REST + WebSocket API
servers:
  - url: http://localhost:8000
    description: Development
security:
  - bearerAuth: []
paths:
...
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, name, role]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
                name: { type: string, minLength: 1, maxLength: 255 }
                role: { type: string, enum: [architect, client] }
                invitation_token: { type: string, description: "Required for client registration" }
      responses:
        "201": { description: User created }
        "400": { description: Validation error or email already registered }
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login and receive tokens
      security: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string, description: "User email" }
                password: { type: string }
      responses:
        "200":
          description: Success
          headers:
            Set-Cookie:
              schema: { type: string }
              description: refresh_token httpOnly cookie
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token: { type: string }
                  token_type: { type: string }
        "401": { description: Invalid credentials }
  /api/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      security: []
      parameters:
        - in: cookie
          name: refresh_token
          schema: { type: string }
          required: true
      responses:
        "200":
          description: New access token
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token: { type: string }
                  token_type: { type: string }
        "401": { description: Invalid or missing refresh token }
...
  /api/auth/forgot-password:
    post:
      tags: [Auth]
      summary: Request password reset email
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        "200": { description: "If email exists, reset link sent" }
  /api/auth/reset-password:
    post:
      tags: [Auth]
      summary: Reset password with token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, new_password]
              properties:
                token: { type: string }
                new_password: { type: string, minLength: 8 }
      responses:
        "200": { description: Password reset successfully }
        "400": { description: Invalid or expired token }
  /api/auth/verify-email/{token}:
    post:
      tags: [Auth]
      summary: Verify email address
      security: []
      parameters:
        - in: path
          name: token
          required: true
          schema: { type: string }
      responses:
        "200": { description: Email verified }
        "400": { description: Invalid or expired token }

  # ── Users ─────────────────────────────────────────────────
...
  /api/files/{file_id}/comments:
    get:
      tags: [Comments]
      summary: List comments on a file (threaded)
      responses:
        "200":
          description: Comment list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CommentResponse"
    post:
      tags: [Comments]
      summary: Add a comment (client or architect)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties:
                body: { type: string, minLength: 1, maxLength: 5000 }
                parent_id: { type: integer, nullable: true, description: "For threaded replies" }
      responses:
        "201":
          description: Comment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommentResponse"
  /api/comments/{comment_id}:
    patch:
      tags: [Comments]
      summary: Edit comment (author only) or mark resolved (architect only)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                body: { type: string, maxLength: 5000 }
                is_resolved: { type: boolean }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommentResponse"
    delete:
      tags: [Comments]
      summary: Delete comment (author or architect only)
      responses:
        "204": { description: Deleted }

  # ── WebSocket ─────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    UserResponse:
      type: object
      properties:
        id: { type: integer }
        email: { type: string }
        name: { type: string }
        role: { type: string, enum: [architect, client] }
        firm_id: { type: integer, nullable: true }
        is_firm_admin: { type: boolean }
        is_verified: { type: boolean }
        created_at: { type: string, format: date-time }
...
    CommentResponse:
      type: object
      properties:
        id: { type: integer }
        file_id: { type: string, format: uuid }
        parent_id: { type: integer, nullable: true }
        body: { type: string }
        is_resolved: { type: boolean }
        author:
          $ref: "#/components/schemas/UserResponse"
        replies:
          type: array
          items:
            $ref: "#/components/schemas/CommentResponse"
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }